Relocate surgeon command
